remove unused models, moved query models to github-models
