+++ src/association.ts
@@ -1,6 +1,6 @@
-import { Association } from "./github-models";
+import { CommentAuthorAssociation } from "./github-models";
 
-export const associations: Association[] = [
+export const associations: CommentAuthorAssociation[] = [
   'NONE',
   'FIRST_TIMER',
   'FIRST_TIME_CONTRIBUTOR',
@@ -10,7 +10,7 @@ export const associations: Association[] = [
   'OWNER'
 ]
 
-export function getAssociationPriority(association: Association): number {
+export function getAssociationPriority(association: CommentAuthorAssociation): number {
   // Note: this will return -1 for any association that is not found.
   // This will prioritise unknown associations lowest.
   return associations.indexOf(association);
