+++ src/github-models.ts
@@ -1,53 +1,12 @@
+export type PullRequestState = 'OPEN' | 'CLOSED' | 'MERGED'
+export type MergeableState = 'MERGEABLE' | 'CONFLICTING' | 'UNKNOWN'
+export type CommentAuthorAssociation = 'MEMBER' | 'OWNER' | 'COLLABORATOR' | 'CONTRIBUTOR' | 'FIRST_TIME_CONTRIBUTOR' | 'FIRST_TIMER' | 'NONE'
+export type PullRequestReviewState = 'PENDING' | 'COMMENTED' | 'APPROVED' | 'CHANGES_REQUESTED' | 'DISMISSED'
 
 export interface PullRequestReference {
   owner: string;
   repo: string;
   number: number;
-}
-
-export interface CheckPullRequest {
-  number: number;
-}
-
-export interface PullRequest {
-  base: {
-    ref: string
-    sha: string
-    repo: {
-      name: string
-    }
-    user: {
-      login: string
-    }
-  }
-  head: {
-    ref: string
-    sha: string
-    repo: {
-      name: string
-    }
-    user: {
-      login: string
-    }
-  }
-  number: number
-  state: 'open' | 'closed'
-  merged: boolean
-  mergeable: boolean,
-  labels: {
-    name: string
-  }[]
-}
-
-export type ReviewState = 'APPROVED' | 'CHANGES_REQUESTED'
-export type Association = 'COLLABORATOR' | 'CONTRIBUTOR' | 'FIRST_TIMER' | 'FIRST_TIME_CONTRIBUTOR' | 'MEMBER' | 'NONE' | 'OWNER'
-export interface Review {
-  submitted_at: string
-  state: ReviewState
-  user: {
-    login: string
-  }
-  author_association: Association
 }
 
 export interface CheckRun {
@@ -58,25 +17,70 @@ export interface CheckRun {
   conclusion: 'success' | 'failure' | 'neutral' | 'cancelled' | 'timed_out' | 'action_required'
 }
 
-export interface BranchProtection {
-  required_status_checks: {
-    strict: boolean
+export interface PullRequestQueryResult {
+  repository: {
+    pullRequest: {
+      number: number,
+      state: PullRequestState,
+      mergeable: MergeableState,
+      potentialMergeCommit: {
+        oid: string
+      },
+      reviews: {
+        nodes: Array<{
+          authorAssociation: CommentAuthorAssociation,
+          author: {
+            login: string
+          }
+          submittedAt: string,
+          state: PullRequestReviewState
+        }>
+      },
+      labels: {
+        nodes: Array<{
+          name: string
+        }>
+      },
+      authorAssociation: CommentAuthorAssociation,
+      baseRef: {
+        repository: {
+          owner: {
+            login: string
+          },
+          name: string
+        },
+        target: {
+          oid: string
+        }
+        name: string
+      },
+      baseRefOid: string,
+      headRef: {
+        repository: {
+          owner: {
+            login: string
+          },
+          name: string
+        },
+        target: {
+          oid: string
+        }
+        name: string
+      },
+      headRefOid: string,
+      repository: {
+        protectedBranches: {
+          nodes: Array<{
+            name: string,
+            hasRestrictedPushes: boolean,
+            hasStrictRequiredStatusChecks: boolean
+          }>
+        }
+      }
+    }
   }
-  enforce_admins: {
-    enabled: boolean
-  }
-  required_pull_request_reviews: {
-    dismiss_stale_reviews: boolean
-    require_code_owner_reviews: boolean
-    required_approving_review_count: number
-  }
-  restrictions: {}
 }
 
-export interface Branch {
-  name: string;
-  commit: {
-    sha: string
-  }
-  protected: boolean
+export type PullRequestInfo = PullRequestQueryResult['repository']['pullRequest'] & {
+  checkRuns: CheckRun[]
 }
