+++ src/pull-request-query.ts
@@ -1,79 +1,6 @@
-import { PullRequestReference, CheckRun } from './github-models';
-import { MergeableState } from './pull-request-query';
+import { PullRequestReference, CheckRun, PullRequestInfo, PullRequestQueryResult } from './github-models';
 import { Context } from "probot";
 import { result } from './utils';
-
-export type PullRequestState = 'OPEN' | 'CLOSED' | 'MERGED'
-export type MergeableState = 'MERGEABLE' | 'CONFLICTING' | 'UNKNOWN'
-export type CommentAuthorAssociation = 'MEMBER' | 'OWNER' | 'COLLABORATOR' | 'CONTRIBUTOR' | 'FIRST_TIME_CONTRIBUTOR' | 'FIRST_TIMER' | 'NONE'
-export type PullRequestReviewState = 'PENDING' | 'COMMENTED' | 'APPROVED' | 'CHANGES_REQUESTED' | 'DISMISSED'
-export interface PullRequestQueryResult {
-  repository: {
-    pullRequest: {
-      number: number,
-      state: PullRequestState,
-      mergeable: MergeableState,
-      potentialMergeCommit: {
-        oid: string
-      },
-      reviews: {
-        nodes: Array<{
-          authorAssociation: CommentAuthorAssociation,
-          author: {
-            login: string
-          }
-          submittedAt: string,
-          state: PullRequestReviewState
-        }>
-      },
-      labels: {
-        nodes: Array<{
-          name: string
-        }>
-      },
-      authorAssociation: CommentAuthorAssociation,
-      baseRef: {
-        repository: {
-          owner: {
-            login: string
-          },
-          name: string
-        },
-        target: {
-          oid: string
-        }
-        name: string
-      },
-      baseRefOid: string,
-      headRef: {
-        repository: {
-          owner: {
-            login: string
-          },
-          name: string
-        },
-        target: {
-          oid: string
-        }
-        name: string
-      },
-      headRefOid: string,
-      repository: {
-        protectedBranches: {
-          nodes: Array<{
-            name: string,
-            hasRestrictedPushes: boolean,
-            hasStrictRequiredStatusChecks: boolean
-          }>
-        }
-      }
-    }
-  }
-}
-
-export type PullRequestInfo = PullRequestQueryResult['repository']['pullRequest'] & {
-  checkRuns: CheckRun[]
-}
 
 export async function queryPullRequest(github: Context['github'], { owner, repo, number }: PullRequestReference): Promise<PullRequestInfo> {
   const response = await github.query(`
